% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANT-functions.R
\name{ANT_calculator}
\alias{ANT_calculator}
\title{Complete Attention systems calculator}
\usage{
ANT_calculator(folder = "NULL", out = "NULL")
}
\arguments{
\item{folder}{specifies the folder with the ANT data which you want to compute over. If unspecified it enters
an interactive interface to select the directory}

\item{out}{specifies where to save results, default is to not save output}
}
\value{
Returns Attention systmes metrics as well as Accuracy, reaction time, and corresponding third of experimetn metrics was computed off of for all subjects in the speciffied folder
Thirds refer to which third of the experimetnt the summariaes were calculated from. 0 refers to using the whole observation
}
\description{
A function that computes all of the ANT metrics for an entire folder of eprime data. Reteurns both  summarries for
thirds of the experiment and for the whole observation (when thirds =0)
}
