% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANT-functions.R
\name{Attention_systems_calculator}
\alias{Attention_systems_calculator}
\title{Attention systems calculator}
\usage{
Attention_systems_calculator(file, thirds = FALSE)
}
\arguments{
\item{file}{specifies the file path to the eprime file of interest}

\item{thirds}{specifies whether the medians should be calculated for eacah third of the observation time separately. Default is FALSE.}
}
\value{
Returns median(s) reaction times for the specified subject as a dataframe
}
\description{
A function that loads eprime data and calculates median response times from ANT data with the option to split data into thirds or take the median over all observations.
}
