% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANT-functions.R
\name{calc_systems}
\alias{calc_systems}
\title{Median response time calcualtor}
\usage{
calc_systems(df, thirds = FALSE)
}
\arguments{
\item{df}{the dataframe containing ANT data for calculating median reaction times}

\item{thirds}{specifies whether the medians should be calculated for eacah third of the observation time separately. Default is FALSE.}
}
\value{
Returns median(s) reaction times for the specified subject as a dataframe
}
\description{
A function that calculates median values of repsonse times from ANT data with the option to split data into thirds or take the median over all observations.
}
