% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANT-functions.R
\name{eprime_to_dataframe}
\alias{eprime_to_dataframe}
\title{An Eprime conversion function}
\usage{
eprime_to_dataframe(wm_path)
}
\arguments{
\item{wm_path}{specifies the file path to the eprime file of interest}
}
\value{
dataframe containing all eprime entries for Accuracy, Reaction time (RT), and Warning type
}
\description{
A function to convert eprime text files to r dataframes
This function converts eprime data into R usable dataframes. It will take in .eprime files, and will return a dataframe.
}
\keyword{dataframe}
\keyword{eprime,}
