% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGT-functions.R
\name{IGT_count_RT_metrics}
\alias{IGT_count_RT_metrics}
\title{IGT counter}
\usage{
IGT_count_RT_metrics(df)
}
\arguments{
\item{df}{dataframe containing the IGT task data}
}
\value{
dataframe containing LTC, IFL, Good, and Score summaries
}
\description{
Function to calculate count based metrics for the IGT task.
}
\keyword{IGT}
\keyword{eprime,}
