% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGT-functions.R
\name{IGT_load}
\alias{IGT_load}
\title{IGT loader}
\usage{
IGT_load(wm_path)
}
\arguments{
\item{wm_path}{specifies the file path to the eprime file of interest}
}
\value{
dataframe containing all eprime entries for Win, Loss, Score, loan, reaction times, sample, and response
}
\description{
Function to load IGT_data and convert necessary columns
}
\keyword{IGT}
\keyword{eprime,}
