% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGT-functions.R
\name{IGT_metrics}
\alias{IGT_metrics}
\title{IGT metrics}
\usage{
IGT_metrics(wm_path)
}
\arguments{
\item{wm_path}{specifies the file path to the eprime file of interest}
}
\value{
dataframe containing LTC, IFL, Good, reaction times, and score by subject and session
}
\description{
Function to load IGT_data and and calculate summary values from the IGT task
}
\keyword{IGT}
\keyword{eprime,}
