% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANT-functions.R
\name{ANT_calculator}
\alias{ANT_calculator}
\title{Complete Attention systems calculator}
\usage{
ANT_calculator(thirds = FALSE, out = "NULL")
}
\arguments{
\item{thirds}{specifies whether the medians should be calculated for eacah third of the observation time separately. Default is FALSE.}

\item{out}{specifies where to save results, default is to not save output}
}
\value{
Returns Attention systmes metrics as well as Accuracy, reaction time, and corresponding third of experimetn metrics was computed off of for all subjects in the speciffied folder
}
\description{
A function that computes all of the ANT metrics for an entire folder of eprime data.
}
